+++ src/features/cart/ui/CartPageFooter.tsx
@@ -4,7 +4,7 @@ import { css } from '@emotion/react';
 import Button from '../../../shared/ui/Button';
 import { useNavigate } from 'react-router';
 import { ROUTES } from '../../../shared/constants/routeConstants';
-import { useSelectedCartContext } from '../../../shared/context/useCartContext';
+import { useSelectedCartContext } from '../../../shared/context/useSelectedCartContext';
 import styled from '@emotion/styled';
 
 const ButtonCSS = css`
